requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Face detection (YOLOv8)
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Face recognition (InsightFace)
insightface>=0.7.3
onnxruntime>=1.15.0

# Additional utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Note: sqlite3 is built-in with Python, no need to install separately 